refactor(bindings): flatten thin binding subdirs to flat modules
Thin bindings hold one file each; the subdir buys nothing. Flatten
``dataset_types/{local_file_chromadb,weaviate_remote}/dataset_type.py``
and ``sources/noop/noop_source.py`` to flat modules at the parent
level. Filenames now describe the binding pair:
``dataset_types/local_file_chromadb.py`` and
``dataset_types/remote_weaviate.py`` (matching the registered NAME,
not the historical directory spelling).

- Registration paths in dataset_types/__init__.py and
  sources/registry.py updated to the flat module paths.
- The one cross-binding import (remote_weaviate.py pulling NoOpSource)
  updated to sources.noop_source.
- The NAME ClassVars on each binding class are unchanged; this is
  purely a file layout change.
