--- backend/syft_space/components/dataset_types/local_file_chromadb/__init__.py
@@ -1,5 +1,0 @@
-"""LocalFile + ChromaDB dataset type binding package."""
-
-from .dataset_type import LocalFileChromaDBDatasetType
-
-__all__ = ["LocalFileChromaDBDatasetType"]
