--- backend/syft_space/components/sources/noop/__init__.py
@@ -1,5 +1,0 @@
-"""No-op source — stub for read-only / externally-managed datasets."""
-
-from syft_space.components.sources.noop.noop_source import NoOpSource
-
-__all__ = ["NoOpSource"]
